refactor: improve code quality and component structure in Backups module
  - Extract inline components to separate files (BackupListItem, DeviceListItem)
  - Remove useless constructors from React class components
  - Fix negated conditions in conditional renders
  - Add Readonly type modifier to component props
  - Add accessibility role to prevent linter warnings
