+++ src/views/Backups/components/BackupsAsFoldersList.tsx
@@ -1,15 +1,10 @@
 import { skinSkeleton } from '../../../app/shared/Skeleton';
 import folderEmptyImage from '../../../assets/icons/light/folder-open.svg';
-import dateService from '../../../app/core/services/date.service';
-import sizeService from '../../../app/drive/services/size.service';
 import { DriveItemData } from '../../../app/drive/types';
 import { useTranslationContext } from '../../../app/i18n/provider/TranslationProvider';
 import Empty from '../../../app/shared/components/Empty/Empty';
 import { List, MenuItemType } from '@internxt/ui';
-import transformItemService from 'app/drive/services/item-transform.service';
-import { items } from '@internxt/lib';
-import { t } from 'i18next';
-import iconService from 'app/drive/services/icon.service';
+import BackupListItem from './BackupListItem';
 
 export default function BackupsAsFoldersList({
   className = '',
@@ -22,7 +17,7 @@ export default function BackupsAsFoldersList({
   onItemClicked,
   onItemSelected,
   onSelectedItemsChanged,
-}: {
+}: Readonly<{
   className?: string;
   contextMenu: Array<MenuItemType<DriveItemData>>;
   currentItems: DriveItemData[];
@@ -33,8 +28,10 @@ export default function BackupsAsFoldersList({
   onItemClicked: (item: DriveItemData) => void;
   onItemSelected: (changes: { device: DriveItemData; isSelected: boolean }[]) => void;
   onSelectedItemsChanged: (changes: { props: DriveItemData; value: boolean }[]) => void;
-}): JSX.Element {
+}>): JSX.Element {
   const { translate } = useTranslationContext();
+
+  const renderBackupListItem = (item: DriveItemData) => <BackupListItem item={item} onItemClicked={onItemClicked} />;
 
   return (
     <div className={`${className} flex min-h-0 grow flex-col`}>
@@ -65,63 +62,7 @@ export default function BackupsAsFoldersList({
           ]}
           items={currentItems}
           isLoading={isLoading}
-          itemComposition={[
-            (item) => {
-              const ItemIconComponent = iconService.getItemIcon(item.isFolder, item.type);
-
-              return (
-                <div className={'file-list-item group'} data-test={`file-list-${item.isFolder ? 'folder' : 'file'}`}>
-                  <div className="flex shrink-0 min-w-[200px] grow items-center pr-3">
-                    {/* ICON */}
-                    <div className="box-content flex items-center pr-4">
-                      <div className="relative flex h-10 w-10 justify-center drop-shadow-soft">
-                        <ItemIconComponent
-                          className="h-full"
-                          data-test={`file-list-${
-                            item.isFolder ? 'folder' : 'file'
-                          }-${transformItemService.getItemPlainNameWithExtension(item)}`}
-                        />
-                      </div>
-                    </div>
-
-                    {/* NAME */}
-                    <div className="flex w-[200px] grow cursor-pointer items-center truncate pr-2">
-                      <button
-                        data-test={`${item.isFolder ? 'folder' : 'file'}-name`}
-                        className="truncate"
-                        title={
-                          transformItemService.getItemPlainNameWithExtension(item) ?? items.getItemDisplayName(item)
-                        }
-                        onClick={
-                          (item.isFolder && !item.deleted) || (!item.isFolder && item.status === 'EXISTS')
-                            ? () => onItemClicked(item)
-                            : undefined
-                        }
-                      >
-                        <p className="truncate">
-                          {transformItemService.getItemPlainNameWithExtension(item) ?? items.getItemDisplayName(item)}
-                        </p>
-                      </button>
-                    </div>
-                  </div>
-
-                  {/* DATE */}
-                  <div className="block shrink-0 w-date items-center whitespace-nowrap">
-                    {dateService.formatDefaultDate(item.updatedAt, t)}
-                  </div>
-
-                  {/* SIZE */}
-                  <div className="w-size shrink-0 items-center whitespace-nowrap">
-                    {sizeService.bytesToString(item.size, false) === '' ? (
-                      <span className="opacity-25">—</span>
-                    ) : (
-                      sizeService.bytesToString(item.size, false)
-                    )}
-                  </div>
-                </div>
-              );
-            },
-          ]}
+          itemComposition={[renderBackupListItem]}
           onClick={(item) => {
             const unselectedDevices = selectedItems.map((deviceSelected) => ({
               device: deviceSelected,
